techpack: display: Trigger panel reset after first kickoff on non-60 Hz
If we're using a refresh rate different from what the bootloader
initializes the display with, a panel reset is necessary for our custom
refresh rate to take effect.

After the first kickoff, request a panel reset from userspace using the
existing ESD recovery mechanism to update the refresh rate. Note that
this is only done if the target refresh rate is *not* 60 Hz because this
is redundant if we don't need a refresh rate update.

Change-Id: I8a5e004d8d70a7521e1a49fe00c33b71aac30613